requirements.txt
fastapi
uvicorn[standard]
requests
pydantic
mysql-connector
customtkinter
python-dotenv
pygame